[EDIT] Moved distance and keypoint files to tools folder.
